--- pkg/microservice/systemconfig/core/handler/connector.go
@@ -1,72 +1,0 @@
-/*
-Copyright 2021 The KodeRover Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-    http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package handler
-
-import (
-	"github.com/gin-gonic/gin"
-
-	"github.com/koderover/zadig/pkg/microservice/systemconfig/core/service"
-	internalhandler "github.com/koderover/zadig/pkg/shared/handler"
-)
-
-func CreateConnector(c *gin.Context) {
-	ctx := internalhandler.NewContext(c)
-	defer func() { internalhandler.JSONResponse(c, ctx) }()
-
-	args := &service.Connector{}
-	if err := c.ShouldBindJSON(args); err != nil {
-		ctx.Err = err
-		return
-	}
-
-	ctx.Err = service.CreateConnector(args, ctx.Logger)
-}
-
-func ListConnectors(c *gin.Context) {
-	ctx := internalhandler.NewContext(c)
-	defer func() { internalhandler.JSONResponse(c, ctx) }()
-
-	ctx.Resp, ctx.Err = service.ListConnectors(ctx.Logger)
-}
-
-func DeleteConnector(c *gin.Context) {
-	ctx := internalhandler.NewContext(c)
-	defer func() { internalhandler.JSONResponse(c, ctx) }()
-
-	ctx.Err = service.DeleteConnector(c.Param("id"), ctx.Logger)
-}
-
-func GetConnector(c *gin.Context) {
-	ctx := internalhandler.NewContext(c)
-	defer func() { internalhandler.JSONResponse(c, ctx) }()
-
-	ctx.Resp, ctx.Err = service.GetConnector(c.Param("id"), ctx.Logger)
-}
-
-func UpdateConnector(c *gin.Context) {
-	ctx := internalhandler.NewContext(c)
-	defer func() { internalhandler.JSONResponse(c, ctx) }()
-
-	args := &service.Connector{}
-	if err := c.ShouldBindJSON(args); err != nil {
-		ctx.Err = err
-		return
-	}
-	args.ID = c.Param("id")
-
-	ctx.Err = service.UpdateConnector(args, ctx.Logger)
-}
